rename reducer to passes
